Added allDates tests; js_only
