+++ assets/prototype/domain/shared/predicates/datetimes/filters/activeOnly/index.test.ts
@@ -1,6 +1,5 @@
 import dates from '../tests/data';
-import activeOnly from '../activeOnly';
-import activeUpcoming from '../activeUpcoming';
+import activeOnly from './index';
 
 describe('activeOnly', () => {
 	it('Should return an array with active-only dates', () => {
@@ -21,17 +20,4 @@ describe('activeOnly', () => {
 		expect(Array.isArray(filteredDates)).toBe(true);
 		expect(filteredDates.length).toBe(0);
 	});
-});
-
-describe.only('activeUpcoming', () => {
-	it('Should return an array of dates which are active OR upcoming', () => {
-		const filteredDates = activeUpcoming(dates);
-
-		expect(Array.isArray(filteredDates)).toBe(true);
-		filteredDates.forEach((date) => {
-			expect(date.isActive || date.isUpcoming).toBe(true);
-		});
-	});
-
-	it('Should return empty array if there are no active dates', () => {});
 });
