+++ assets/prototype/domain/shared/predicates/datetimes/filters/activeOnly/index.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { Datetime } from '../../../../eventEditor/data/types';
+import { Datetime } from '../../../../../eventEditor/data/types';
 
 const activeOnly = (dates: Datetime[]): Datetime[] | [] => dates.filter(({ isActive }) => isActive);
