+++ assets/prototype/domain/shared/predicates/datetimes/filters/activeUpcoming/index.test.ts
@@ -1,37 +1,15 @@
-/**
- * Internal dependencies
- */
-import aboveCapacity from '../aboveCapacity/index';
 import dates from '../tests/data';
+import activeUpcoming from '.';
 
-test('Should filter out dates based on above50Capacity show type', () => {
-	const filteredDates = aboveCapacity({ capacity: 50, dates });
+describe.only('activeUpcoming', () => {
+	it('Should return an array of dates which are active OR upcoming', () => {
+		const filteredDates = activeUpcoming(dates);
 
-	expect(filteredDates.length).toBe(6);
-	expect(filteredDates[0].capacity).toBe(10);
-	expect(filteredDates[0].sold).toBe(5);
-	expect(filteredDates[5].capacity).toBe(10);
-	expect(filteredDates[5].sold).toBe(10);
-});
+		expect(Array.isArray(filteredDates)).toBe(true);
+		filteredDates.forEach((date) => {
+			expect(date.isActive || date.isUpcoming).toBe(true);
+		});
+	});
 
-test('Should filter out dates based on above75Capacity show type', () => {
-	const filteredDates = aboveCapacity({ capacity: 75, dates });
-
-	expect(filteredDates.length).toBe(3);
-	expect(filteredDates[0].capacity).toBe(10);
-	expect(filteredDates[0].sold).toBe(8);
-	expect(filteredDates[1].capacity).toBe(10);
-	expect(filteredDates[1].sold).toBe(9);
-	expect(filteredDates[2].capacity).toBe(10);
-	expect(filteredDates[2].sold).toBe(10);
-});
-
-test('Should filter out dates based on above90Capacity show type', () => {
-	const filteredDates = aboveCapacity({ capacity: 90, dates });
-
-	expect(filteredDates.length).toBe(2);
-	expect(filteredDates[0].capacity).toBe(10);
-	expect(filteredDates[0].sold).toBe(9);
-	expect(filteredDates[1].capacity).toBe(10);
-	expect(filteredDates[1].sold).toBe(10);
+	it('Should return empty array if there are no active dates', () => {});
 });
