--- assets/prototype/domain/shared/predicates/datetimes/filters/allDates.ts
@@ -1,11 +1,0 @@
-/**
- * Internal dependencies
- */
-import { Datetime } from '../../../../eventEditor/data/types';
-
-const allDates = (dates: Datetime[]) => {
-	const withoutTrashed = (date) => date.deleted;
-	return dates.filter(withoutTrashed);
-};
-
-export default allDates;
